% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_multi_col_freqs.R
\name{calc_multi_col_freqs}
\alias{calc_multi_col_freqs}
\title{Calculate multiple column frequency table}
\usage{
calc_multi_col_freqs(data, levels, labels, calc_props = FALSE)
}
\arguments{
\item{data}{the selected columns from the CARS-3 wave dataset}

\item{levels}{a list of factor levels for the cell values of all columns}

\item{labels}{a list of labels for each of the variables}

\item{calc_props}{whether or not to calculate proportions (logical, FALSE by default)}
}
\value{
frequencies data frame with the frequencies of response options by column
}
\description{
Take multiple columns and produce a frequency or proportion of each column
}
