% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsurvey3-package.R
\docType{package}
\name{carsurvey3}
\alias{carsurvey3}
\title{carsurvey3}
\description{
Analysis for the Coding in Analysis and Research Survey (CARS) wave 2 (Sep-Oct 2021).
The package produces frequencies and outputs for CARS data.
}
