% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars.R
\name{derive_rap_score}
\alias{derive_rap_score}
\title{Derive RAP scores}
\usage{
derive_rap_score(data)
}
\arguments{
\item{data}{a date frame containing cleaned CARS wave 3 data}
}
\value{
dataframe containing the additional RAP score columns
}
\description{
Derive basic and advanced RAP score columns from existing variables and add to the dataframe.
}
