% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars.R
\name{derive_vars}
\alias{derive_vars}
\title{Derive variables}
\usage{
derive_vars(data)
}
\arguments{
\item{data}{tidied and relabelled CARS wave 3 dataset.}
}
\value{
data (data.frame).
}
\description{
API function for deriving additional variables.
}
