% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{Dummy dataset documentation}
\description{
this is a dummy dataset based on the original CARS dataset. This is NOT synthetic data and should only be used for testing.
}
