% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_code_freq_streaming}
\alias{enforce_code_freq_streaming}
\title{Enforce code frequency streaming}
\usage{
enforce_code_freq_streaming(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then multiple questions are skipped
}
