% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_codefreq_ci}
\alias{enforce_codefreq_ci}
\title{Enforce code frequency streaming on continuous integration questions}
\usage{
enforce_codefreq_ci(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then continuous integration questions is skipped
}
