% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_codefreq_coding_prac}
\alias{enforce_codefreq_coding_prac}
\title{Enforce code frequency streaming on coding practices}
\usage{
enforce_codefreq_coding_prac(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then coding practices is skipped
}
