% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_codefreq_comment}
\alias{enforce_codefreq_comment}
\title{Enforce code frequency streaming on extra comment question}
\usage{
enforce_codefreq_comment(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then extra comment question is skipped
}
