% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_codefreq_dep_man}
\alias{enforce_codefreq_dep_man}
\title{Enforce code frequency streaming on dependency management questions}
\usage{
enforce_codefreq_dep_man(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then dependency management questions is skipped
}
