% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_codefreq_doc}
\alias{enforce_codefreq_doc}
\title{Enforce code frequency streaming on documentation questions}
\usage{
enforce_codefreq_doc(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then documentation practices is skipped
}
