% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_codefreq_rep_wf}
\alias{enforce_codefreq_rep_wf}
\title{Enforce code frequency streaming on reproducible workflow questions}
\usage{
enforce_codefreq_rep_wf(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if code frequency is never then reproducible workflow questions is skipped
}
