% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_degree_streaming}
\alias{enforce_degree_streaming}
\title{Enforce degree streaming}
\usage{
enforce_degree_streaming(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if highest qualification isn't degree then no degree subject has "Yes" returned
}
