% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_heard_of_rap_streaming}
\alias{enforce_heard_of_rap_streaming}
\title{Enforce heard of RAP streaming}
\usage{
enforce_heard_of_rap_streaming(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if heard of RAP is "No" then skip various RAP based questions
}
