% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_prev_exp_streaming}
\alias{enforce_prev_exp_streaming}
\title{Enforce previous coding experience streaming}
\usage{
enforce_prev_exp_streaming(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if previous coding experience is no then where learned to code is skipped
}
