% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_rap_champ_knowl_streaming}
\alias{enforce_rap_champ_knowl_streaming}
\title{Enforce heard of RAP streaming on RAP_champion and RAP_knowledge}
\usage{
enforce_rap_champ_knowl_streaming(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if heard of RAP is "No" then are you RAP_champion and RAP_knowledge are skipped
}
