% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_rap_champ_streaming}
\alias{enforce_rap_champ_streaming}
\title{Enforce RAP champion streaming}
\usage{
enforce_rap_champ_streaming(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data frame
}
\description{
Enforce the streaming rules that if RAP champion is yes then Do you know your RAP champion is skipped
}
