% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_streaming.R
\name{enforce_streaming}
\alias{enforce_streaming}
\title{API to enforce streaming rules}
\usage{
enforce_streaming(data)
}
\arguments{
\item{data}{pre-processed data}
}
\value{
data
}
\description{
Enforce all streaming rules across all questions
}
