% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{freq_subplots}
\alias{freq_subplots}
\title{Plot frequency subplots}
\usage{
freq_subplots(
  data,
  xlab,
  ylab,
  height,
  width,
  bar_colour,
  nrows = 3,
  y_margin = 0.1,
  x_margin = 0.1,
  font_size = 12,
  orientation = "h"
)
}
\arguments{
\item{data}{Frequency table (data frame). Expects a tidy dataset with 3 columns: factor 1 (categorical axes), factor 2 (one plot per factor level) and values.}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{height}{plot height in pixels}

\item{width}{plot width in pixels}

\item{bar_colour}{Colour name. Defaults to blue (see carsurvey2::get_gradient())}

\item{nrows}{number of rows in plot (min 2, 3 by default)}

\item{y_margin}{margin between rows (chart proportion), 0.1 by default.}

\item{x_margin}{margin between columns (chart proportion). 0.1 by default.}

\item{font_size}{minimum font size for the plot (numeric).}

\item{orientation}{plot orientation ("h" = horizontal, "v" = verical). Vertical by default}
}
\value{
subplots
}
\description{
Produce multiple bar chart (plotly) for single factor frequency data.
Intended for use with frequencies represented as percentages.
Incomplete validation checks: some combinations of number of plots and number of rows/columns will fail:
Each column should contain at least two charts.
}
