% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tables.R
\name{plot_factorial}
\alias{plot_factorial}
\title{Plot factorial frequency graph}
\usage{
plot_factorial(table, xlab, ylab, n, font_size = 12, orientation = "v", ...)
}
\arguments{
\item{table}{Frequency table (data frame). 3 columns - cateogry names, groups and frequencies.}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{n}{sample size}

\item{font_size}{minimum font size for the plot (numeric).}

\item{orientation}{plot orientation ("h" = horizontal, "v" = verical). Vertical by default.}

\item{...}{additional plotly_ly arguments}
}
\value{
bar chart
}
\description{
Produce bar chart (plotly) for frequency data with grouping variable.
}
