% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_stacked}
\alias{plot_stacked}
\title{Plot stacked bar graph}
\usage{
plot_stacked(
  table,
  xlab,
  ylab,
  n,
  colour_scale = "2gradients",
  font_size = 12,
  neutral_mid = TRUE,
  ...
)
}
\arguments{
\item{table}{Frequency table for stacked bar chart (data frame). 3+ columns - sub-question names in column 1 with answer options in subsequent columns..}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{n}{sample size}

\item{colour_scale}{type of colour scale ("gradient", "scale" or "2gradients"). See get_gradient(), get_2colour_scale() and get_2colour_gradients().}

\item{font_size}{minimum font size for the plot (numeric).}

\item{neutral_mid}{whether the midpoint of the colour scale should be neutral ("2gradients" scale only). TRUE by default}

\item{...}{additional plotly_ly arguments}
}
\value{
bar chart
}
\description{
Produce stacked bar chart (plotly).
}
