% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sub_freqs}
\alias{plot_sub_freqs}
\title{Plot frequency graph for subplot}
\usage{
plot_sub_freqs(
  table,
  xlab,
  ylab,
  height,
  width,
  title = "",
  bar_colour,
  font_size = 12,
  orientation = "v"
)
}
\arguments{
\item{table}{Frequency table (data frame). 2 columns - cateogry names and frequencies.}

\item{xlab}{X axis title}

\item{ylab}{Y axis title}

\item{height}{plot height in pixels}

\item{width}{plot width in pixels}

\item{title}{optional. Subplot title}

\item{bar_colour}{Colour name. Defaults to blue (see carsurvey2::get_gradient())}

\item{font_size}{minimum font size for the plot (numeric).}

\item{orientation}{plot orientation ("h" = horizontal, "v" = verical). Vertical by default}
}
\value{
bar chart
}
\description{
Produce bar chart (plotly) for single factor frequency data.
}
