% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_grade.R
\name{recode_grade}
\alias{recode_grade}
\title{Recode grade}
\usage{
recode_grade(data, grade_col = "grade", dep_col = "department")
}
\arguments{
\item{data}{a data frame containing cleaned CARS wave 2 data}

\item{grade_col}{the name of the column containing the grade (set to "grade" by default)}

\item{dep_col}{the name of the column containing the department (set to "dept" by default)}
}
\value{
cARS wave 3 data with the recoded grade column
}
\description{
Recode non-set responses to grade question
}
