% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols.R
\name{rename_cols}
\alias{rename_cols}
\title{Rename columns}
\usage{
rename_cols(data)
}
\arguments{
\item{data}{CARS wave 3 (2021) survey data (data.frame).}
}
\value{
data.frame
}
\description{
add meaningful column names to dataset ingested from smartsurvey API.
}
