% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_main_site.R
\name{render_main_site}
\alias{render_main_site}
\title{Render main site}
\usage{
render_main_site(data, markdown_file_path = "rmarkdown/main")
}
\arguments{
\item{data}{This is generated using the carsurvey2::data_ functions.}

\item{markdown_file_path}{the path containing the rmarkdown site documents}
}
\description{
Processes the Smart survey data to generate a series of tables.
Then removes the old site and uses the generated tables to render a new site.
Uses the markdown files located in the internal var markdown_file_path.
}
