% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_prof_pages}
\alias{render_prof_pages}
\title{Render profession pages}
\usage{
render_prof_pages(
  data,
  page_title = "",
  output_folder = "docs",
  template_path = "rmarkdown/summary_template/summary.rmd"
)
}
\arguments{
\item{data}{CARS wave 3 dataset.}

\item{page_title}{This is ONLY PART of the title. The title is generated from "DRAFT: ", page_title , " profile: ", filter --
where filter is one element of the list as described above in filter_variable (values over 20) and the page title is this argument.}

\item{output_folder}{Folder the site is built and saved to}

\item{template_path}{The path to the template that gets render for each department}
}
\description{
Creates pages by multi-column filter (profession)
}
