% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_ability_change}
\alias{summarise_ability_change}
\title{Summarise ability change frequency}
\usage{
summarise_ability_change(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for ability change
}
