% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_adv_score_by_imp}
\alias{summarise_adv_score_by_imp}
\title{Compare advanced RAP score to implementation of RAP}
\usage{
summarise_adv_score_by_imp(implementing_data)
}
\arguments{
\item{implementing_data}{carsurvey data filtered to people who have heard of RAP and code at least rarely}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for advanced rap score compared with implementation of RAP
}
