% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_adv_score_by_understanding}
\alias{summarise_adv_score_by_understanding}
\title{Compare advanced RAP score to understanding of key RAP components}
\usage{
summarise_adv_score_by_understanding(implementing_data)
}
\arguments{
\item{implementing_data}{carsurvey data filtered to people who have heard of RAP and code at least rarely}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for advanced rap score compared with understanding of key RAP components
}
