% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_all}
\alias{summarise_all}
\title{Produce all summary tables}
\usage{
summarise_all(data, all_tables = FALSE)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}

\item{all_tables}{logical: whether to produce all summary output tables. Defaults to FALSE.}
}
\value{
list of frequency tables
}
\description{
Produce all summary tables and return as a named list.
}
