% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_all_further}
\alias{summarise_all_further}
\title{Produce all summary tables for further analysis}
\usage{
summarise_all_further(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
list of frequency tables
}
\description{
Produce all summary tables for further analysis and return as a named list.
}
