% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_basic_score_by_imp}
\alias{summarise_basic_score_by_imp}
\title{Compare basic RAP score to implementation of RAP}
\usage{
summarise_basic_score_by_imp(implementing_data)
}
\arguments{
\item{implementing_data}{carsurvey data filter to people who have heard of RAP and code at least rarely}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for basic rap score compared with implementation of RAP
}
