% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_code_freq}
\alias{summarise_code_freq}
\title{Summarise coding frequency}
\usage{
summarise_code_freq(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{

}
\description{
calculate frequency table for coding frequency.
}
