% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_coding_tools}
\alias{summarise_coding_tools}
\title{Summarise coding tools}
\usage{
summarise_coding_tools(data, type = list("knowledge", "access"))
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}

\item{type}{type of table (knowledge or access)}
}
\value{

}
\description{
calculate frequency table coding tools (knowledge or access)
}
