% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_coding_tools_comparison}
\alias{summarise_coding_tools_comparison}
\title{Compare coding tool frequency of ONS to other departments}
\usage{
summarise_coding_tools_comparison(
  ons_data,
  other_deps_data,
  ons_tables,
  other_deps_tables
)
}
\arguments{
\item{ons_data}{carsurvey data filter to ONS department}

\item{other_deps_data}{lcarsurvey  data filter to all other department}

\item{ons_tables}{list of tables made from summaries_all from carsurvey data filter to ONS department}

\item{other_deps_tables}{list of tables made from summaries_all from carsurvey  data filter to all other department}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for coding tool for ONS and all other departments combined
}
