% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_doc}
\alias{summarise_doc}
\title{Frequency of documentation use}
\usage{
summarise_doc(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
frequency table (data.frame)
}
\description{
Create frequency table of documentation use
}
