% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_language_status}
\alias{summarise_language_status}
\title{Summarise programming language status}
\usage{
summarise_language_status(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
frequency table (data.frame)
}
\description{
calculate counts of responents reporting access to, knowledge of, or both for each programming language.
}
