% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_languages_by_prof}
\alias{summarise_languages_by_prof}
\title{Summarise programming language knowledge by profession}
\usage{
summarise_languages_by_prof(data)
}
\arguments{
\item{data}{CARS data (preprocessed)}
}
\value{
data.frame
}
\description{
only used the main summary page. Needs to be turned into wide data for html table.
}
