% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_operations}
\alias{summarise_operations}
\title{Summarise data operations}
\usage{
summarise_operations(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for data operations
}
