% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_rap_basic}
\alias{summarise_rap_basic}
\title{Summarise basic rap score}
\usage{
summarise_rap_basic(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for basic rap scores
}
