% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_rap_comp}
\alias{summarise_rap_comp}
\title{RAP score components}
\usage{
summarise_rap_comp(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
frequency table (data.frame)
}
\description{
Create frequency table of basic and advanced RAP score components
}
