% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_rap_opinions}
\alias{summarise_rap_opinions}
\title{Opinions of RAP}
\usage{
summarise_rap_opinions(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
frequency table (data.frame)
}
\description{
Create frequency table of opinions of RAP
}
