% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_where_learned_code}
\alias{summarise_where_learned_code}
\title{Summarise where respondents learned to code}
\usage{
summarise_where_learned_code(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table of where respondents learned to code
}
