% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_correlation_test}
\alias{compute_correlation_test}
\title{compute spearmans rank test}
\usage{
compute_correlation_test(first_variable, second_variable, method = "spearman")
}
\arguments{
\item{first_variable}{first variable to compare}

\item{second_variable}{second variable to compare}

\item{method}{optional: correlation method. Defaults to "spearman"}
}
\value{
list of correlation test outputs
}
\description{
compute spearmans rank correlation on two variables
}
