% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_language_status.R
\name{derive_language_status}
\alias{derive_language_status}
\title{Derive code status}
\usage{
derive_language_status(data)
}
\arguments{
\item{data}{tidied CARS wave 3 data (data.frame).}
}
\value{
data.frame
}
\description{
Dervice the staus of each programmming language as "access" (access only), "knowledge" (knowledge only), "both" or "neither".
}
