% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_all}
\alias{summarise_all}
\title{Produce all summary tables}
\usage{
summarise_all(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
list of frequency tables
}
\description{
Produce all summary tables and return as a named list.
}
