% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{summarise_cap_change_by_freq}
\alias{summarise_cap_change_by_freq}
\title{Summarise manage someone who codes}
\usage{
summarise_cap_change_by_freq(data)
}
\arguments{
\item{data}{full CARS wave 3 data.frame after preprocessing}
}
\value{
frequency table (data.frame)
}
\description{
calculate frequency table for if someone line manages someone who codes
}
